add log verbose
