+++ bin/run
@@ -1,26 +1,9 @@
 #!/usr/bin/env node
 
-const { Command } = require('commander');
-
-const packageJson = require('../package.json');
-
-function increaseVerbosity(_, previous) {
-  return previous + 1;
-}
-
-const program = new Command();
-program
-.version(packageJson.version)
-.option('-f, --force', 'force to generate output json files')
-.option('-v, --verbose', 'verbosity that can be increased', increaseVerbosity, 0)
-.parse(process.argv);
-
 const theApp = require('../lib');
-theApp.setLoggerLevel(program.verbose);
 
 new Promise(async (resolve, reject) => {
   // do a thing, possibly async, then…
-  theApp.logger.debug('verbose', program.verbose);
   const retCode =  await theApp.run();
   if (0 === retCode) {
     resolve(0);
