+++ lib/app/inquirer.js
@@ -3,47 +3,25 @@ const path = require('path');
 
 const YAML = require('yaml');
 const { prompt } = require('enquirer');
-const Logger = require('js-logger');
+
 
 const  { ReturnCode } = require('../common/return-code');
-
-// Predefined logging levels.
-// Logger.TRACE = defineLogLevel(1, 'TRACE');
-// Logger.DEBUG = defineLogLevel(2, 'DEBUG');
-// Logger.INFO = defineLogLevel(3, 'INFO');
-// Logger.TIME = defineLogLevel(4, 'TIME');
-// Logger.WARN = defineLogLevel(5, 'WARN');
-// Logger.ERROR = defineLogLevel(8, 'ERROR');
-const LogLevel = [  
-  Logger.ERROR,
-  Logger.WARN,
-  Logger.TIME,
-  Logger.INFO,
-  Logger.DEBUG,
-  Logger.TRACE,
-];
 
 const configFilePath = './inquirer.yaml';
 const outputBasePath = './.inquirer';
 class Inquirer {
-  constructor() {
+  constructor(program, Logger) {
+    this.program = program;
     this.logger = Logger;
-    this.logger.useDefaults();
-    this.logger.setLevel(this.logger.OFF);
+    this.setLoggerLevel(program.verbose);
+
+    this.logger.debug('verbose', program.verbose);
 
     fs.mkdirSync(outputBasePath, { recursive: true });
   }
 
   setLoggerLevel(level) {
-    if (!level) {
-      return;
-    }
-    if (level > LogLevel.length) {
-      level = LogLevel.length - 1;
-    }
-
-    const theLevel = LogLevel[level];
-    this.logger.setLevel(theLevel);
+    this.logger.setLoggerLevel(level)
   }
 
   async run() {
